Refactored some specs, file names etc,
